pyproject.toml
[project]
name = "autolog"
version = "0.1.0"
description = "AI-powered structured project logging for researchers and engineers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "chromadb>=1.5.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "torch",
    "torchvision",
    "transformers>=5.3.0",
    "typer>=0.12.0",
]

[project.scripts]
autolog = "autolog.cli.main:app"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]
